requirements.txt
fastapi
uvicorn
python-multipart
openpyxl
pymupdf
pillow